+++ Core.cpk/Scripts/Items/Equipment/Braided/ItemBraidedPants.cs
@@ -4,7 +4,7 @@
     {
         public override string Description => GetProtoEntity<ItemBraidedChestplate>().Description;
 
-        public override ushort DurabilityMax => 500;
+        public override uint DurabilityMax => 500;
 
         public override string Name => "Braided pants";
