+++ Core.cpk/Scripts/Items/Implants/ItemImplantBroken.cs
@@ -11,7 +11,7 @@
         public override string Description =>
             "You have leftover pieces of a broken implant in your body. It doesn't do you any good and you should remove it as soon as possible.";
 
-        public override ushort DurabilityMax => 0; // non-degradeable
+        public override uint DurabilityMax => 0; // non-degradeable
 
         public override TimeSpan Lifetime => TimeSpan.Zero; // non-degradeable
