+++ Core.cpk/Scripts/Items/Weapons/Melee/ItemNoWeapon.cs
@@ -19,7 +19,7 @@
         [NotLocalizable]
         public override string Description => "Fallback weapon prototype in case no weapon selected.";
 
-        public override ushort DurabilityMax => 0;
+        public override uint DurabilityMax => 0;
 
         public override ITextureResource Icon => null;
