+++ Core.cpk/Scripts/Items/Weapons/MobWeapons/ItemWeaponGenericAnimalStrong.cs
@@ -20,7 +20,7 @@
 
         public override string Description => null;
 
-        public override ushort DurabilityMax => 0;
+        public override uint DurabilityMax => 0;
 
         public override double FireInterval => 1.5;
