+++ Core.cpk/Scripts/Items/Weapons/MobWeapons/ItemWeaponSnakeBiteStrong.cs
@@ -19,7 +19,7 @@
 
         public override string Description => null;
 
-        public override ushort DurabilityMax => 0;
+        public override uint DurabilityMax => 0;
 
         public override double FireAnimationDuration => 0.5;
