+++ Core.cpk/Scripts/Items/Weapons/Ranged/ItemHandgun10mm.cs
@@ -21,7 +21,7 @@
 
         public override string Description => "Handgun developed for 10mm rounds.";
 
-        public override ushort DurabilityMax => 250;
+        public override uint DurabilityMax => 250;
 
         public override string Name => "Handgun";
